[Add] SideMenu component
